+++ packages/toolkit/config/paths.config.js
@@ -1,6 +1,7 @@
 module.exports = {
-	srcDir: './assets/',
-	cssLoaderPaths: ['./assets/css', './includes/blocks'],
-	copyAssetsDir: './assets/',
-	blocksDir: './includes/blocks/',
+	srcDir: "./assets/",
+	cssLoaderPaths: ["./assets/css", "./includes/blocks"],
+	copyAssetsDir: "./assets/",
+	blocksDir: "./includes/blocks/",
+	iconsDir: "./assets/icons/",
 };