+++ packages/toolkit/config/webpack/plugins.js
@@ -1,18 +1,19 @@
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-const ESLintPlugin = require('eslint-webpack-plugin');
-const DependencyExtractionWebpackPlugin = require('@wordpress/dependency-extraction-webpack-plugin');
-const MiniCSSExtractPlugin = require('mini-css-extract-plugin');
-const StyleLintPlugin = require('stylelint-webpack-plugin');
-const WebpackBar = require('webpackbar');
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const { BundleAnalyzerPlugin } = require('webpack-bundle-analyzer');
-const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-const { resolve } = require('path');
-const RemoveEmptyScriptsPlugin = require('./plugins/remove-empty-scripts');
-const CleanExtractedDeps = require('./plugins/clean-extracted-deps');
-const WDCToolkitTscPlugin = require('./plugins/tsc');
-const NoBrowserSyncPlugin = require('./plugins/no-browser-sync');
+const CopyWebpackPlugin = require("copy-webpack-plugin");
+const ESLintPlugin = require("eslint-webpack-plugin");
+const DependencyExtractionWebpackPlugin = require("@wordpress/dependency-extraction-webpack-plugin");
+const MiniCSSExtractPlugin = require("mini-css-extract-plugin");
+const StyleLintPlugin = require("stylelint-webpack-plugin");
+const WebpackBar = require("webpackbar");
+const path = require("path");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
+const ReactRefreshWebpackPlugin = require("@pmmmwh/react-refresh-webpack-plugin");
+const { resolve } = require("path");
+const RemoveEmptyScriptsPlugin = require("./plugins/remove-empty-scripts");
+const CleanExtractedDeps = require("./plugins/clean-extracted-deps");
+const WDCToolkitTscPlugin = require("./plugins/tsc");
+const NoBrowserSyncPlugin = require("./plugins/no-browser-sync");
+const SpriteLoaderPlugin = require("svg-sprite-loader/plugin");
 
 const {
 	hasStylelintConfig,
@@ -20,17 +21,19 @@ const {
 	hasProjectFile,
 	getArgFromCLI,
 	maybeInsertStyleVersionHash,
-} = require('../../utils');
-const { isPackageInstalled } = require('../../utils/package');
+} = require("../../utils");
+const { isPackageInstalled } = require("../../utils/package");
 
 const removeDistFolder = (file) => {
-	return file.replace(/(^\.\/dist\/)|^dist\//, '');
+	return file.replace(/(^\.\/dist\/)|^dist\//, "");
 };
 
 // There are differences between Windows and Posix when it comes to the WebpackBar
 // This ensures that the same reporter is used everywhere
 const webpackbarArguments =
-	process.env.JEST_WORKER_ID !== undefined ? { reporters: ['basic'] } : undefined;
+	process.env.JEST_WORKER_ID !== undefined
+		? { reporters: ["basic"] }
+		: undefined;
 
 module.exports = ({
 	isPackage,
@@ -52,27 +55,29 @@ module.exports = ({
 	const hasReactFastRefresh = hot && !isProduction;
 
 	const hasBrowserSync =
-		isPackageInstalled('browser-sync-webpack-plugin') && isPackageInstalled('browser-sync');
+		isPackageInstalled("browser-sync-webpack-plugin") &&
+		isPackageInstalled("browser-sync");
 
-	const shouldLoadBrowserSync = !isProduction && devURL && !hasReactFastRefresh && hasBrowserSync;
+	const shouldLoadBrowserSync =
+		!isProduction && devURL && !hasReactFastRefresh && hasBrowserSync;
 
 	let browserSync = !isProduction && devURL ? new NoBrowserSyncPlugin() : false;
 	if (shouldLoadBrowserSync) {
 		// eslint-disable-next-line global-require, import/no-extraneous-dependencies
-		const BrowserSyncPlugin = require('browser-sync-webpack-plugin');
+		const BrowserSyncPlugin = require("browser-sync-webpack-plugin");
 		browserSync = new BrowserSyncPlugin(
 			{
-				host: 'localhost',
-				port: getArgFromCLI('--port') || 3000,
+				host: "localhost",
+				port: getArgFromCLI("--port") || 3000,
 				proxy: devURL,
 				open: false,
-				files: ['**/*.php', '**/*.js', 'dist/**/*.css'],
-				ignore: ['dist/**/*.php', 'dist/**/*.js'],
-				serveStatic: ['.'],
+				files: ["**/*.php", "**/*.js", "dist/**/*.css"],
+				ignore: ["dist/**/*.php", "dist/**/*.js"],
+				serveStatic: ["."],
 				rewriteRules: [
 					{
 						match: /wp-content\/themes\/.*\/dist/g,
-						replace: 'dist',
+						replace: "dist",
 					},
 				],
 			},
@@ -87,9 +92,11 @@ module.exports = ({
 
 	return [
 		devServer &&
-		new HtmlWebpackPlugin({
-			...(hasProjectFile('public/index.html') && { template: 'public/index.html' }),
-		}),
+			new HtmlWebpackPlugin({
+				...(hasProjectFile("public/index.html") && {
+					template: "public/index.html",
+				}),
+			}),
 		new ESLintPlugin({
 			failOnError: false,
 			fix: false,
@@ -125,9 +132,9 @@ module.exports = ({
 
 					return fullPath
 						? !path
-							.relative(blocksSourceDirectory, fullPath)
-							// startWith('../') but in a cross-env way
-							.startsWith(path.join('..', '/'))
+								.relative(blocksSourceDirectory, fullPath)
+								// startWith('../') but in a cross-env way
+								.startsWith(path.join("..", "/"))
 						: false;
 				});
 
@@ -144,52 +151,56 @@ module.exports = ({
 
 				return isBlockAsset ? filenames.blockCSS : filenames.css;
 			},
-			chunkFilename: '[id].css',
+			chunkFilename: "[id].css",
 		}),
 
 		!isPackage &&
-		// Copy static assets to the `dist` folder.
-		new CopyWebpackPlugin({
-			patterns: [
-				{
-					from: '**/*.{jpg,jpeg,png,gif,webp,avif,ico,svg,eot,ttf,woff,woff2,otf}',
-					to: '[path][name][ext]',
-					noErrorOnMissing: true,
-					context: path.resolve(process.cwd(), paths.copyAssetsDir),
-				},
-				useBlockAssets && {
-					from: path.join(blocksSourceDirectory, '**/block.json').replace(/\\/g, '/'),
-					context: blocksSourceDirectory,
-					noErrorOnMissing: true,
-					to: 'blocks/[path][name][ext]',
-					transform: (content, absoluteFilename) => {
-						return maybeInsertStyleVersionHash(content, absoluteFilename);
+			// Copy static assets to the `dist` folder.
+			new CopyWebpackPlugin({
+				patterns: [
+					{
+						from: "**/*.{jpg,jpeg,png,gif,webp,avif,ico,svg,eot,ttf,woff,woff2,otf}",
+						to: "[path][name][ext]",
+						noErrorOnMissing: true,
+						context: path.resolve(process.cwd(), paths.copyAssetsDir),
 					},
-				},
-				useBlockAssets && {
-					from: path.join(blocksSourceDirectory, '**/*.php').replace(/\\/g, '/'),
-					context: blocksSourceDirectory,
-					noErrorOnMissing: true,
-					to: 'blocks/[path][name][ext]',
-				},
-				hasReactFastRefresh && {
-					from: fromConfigRoot('fast-refresh.php'),
-					to: '[path][name][ext]',
-					noErrorOnMissing: true,
-					context: path.resolve(process.cwd(), '/dist'),
-				},
-			].filter(Boolean),
-		}),
+					useBlockAssets && {
+						from: path
+							.join(blocksSourceDirectory, "**/block.json")
+							.replace(/\\/g, "/"),
+						context: blocksSourceDirectory,
+						noErrorOnMissing: true,
+						to: "blocks/[path][name][ext]",
+						transform: (content, absoluteFilename) => {
+							return maybeInsertStyleVersionHash(content, absoluteFilename);
+						},
+					},
+					useBlockAssets && {
+						from: path
+							.join(blocksSourceDirectory, "**/*.php")
+							.replace(/\\/g, "/"),
+						context: blocksSourceDirectory,
+						noErrorOnMissing: true,
+						to: "blocks/[path][name][ext]",
+					},
+					hasReactFastRefresh && {
+						from: fromConfigRoot("fast-refresh.php"),
+						to: "[path][name][ext]",
+						noErrorOnMissing: true,
+						context: path.resolve(process.cwd(), "/dist"),
+					},
+				].filter(Boolean),
+			}),
 		devURL && browserSync,
 		// Lint CSS.
 		new StyleLintPlugin({
 			context: path.resolve(process.cwd(), paths.srcDir),
-			files: '**/*.(s(c|a)ss|css)',
+			files: "**/*.(s(c|a)ss|css)",
 			allowEmptyInput: true,
 			lintDirtyModulesOnly: true,
 			failOnError: false,
 			...(!hasStylelintConfig() && {
-				configFile: fromConfigRoot('stylelint.config.js'),
+				configFile: fromConfigRoot("stylelint.config.js"),
 			}),
 		}),
 		// Fancy WebpackBar.
@@ -197,25 +208,32 @@ module.exports = ({
 		// dependencyExternals variable controls whether scripts' assets get
 		// generated, and the default externals set.
 		wpDependencyExternals &&
-		!isPackage &&
-		new DependencyExtractionWebpackPlugin({
-			injectPolyfill: false,
-			requestToHandle: (request) => {
-				if (request.includes('react-refresh/runtime')) {
-					return 'wdc-toolkit-react-refresh-runtime';
-				}
+			!isPackage &&
+			new DependencyExtractionWebpackPlugin({
+				injectPolyfill: false,
+				requestToHandle: (request) => {
+					if (request.includes("react-refresh/runtime")) {
+						return "wdc-toolkit-react-refresh-runtime";
+					}
 
-				return undefined;
-			},
-		}),
+					return undefined;
+				},
+			}),
 		new CleanExtractedDeps(),
 		new RemoveEmptyScriptsPlugin(),
 		new WDCToolkitTscPlugin(),
-		analyze && isProduction && new BundleAnalyzerPlugin({ analyzerMode: 'static' }),
+		analyze &&
+			isProduction &&
+			new BundleAnalyzerPlugin({ analyzerMode: "static" }),
 		hasReactFastRefresh &&
-		new ReactRefreshWebpackPlugin({
-			overlay: { sockHost: '127.0.0.1', sockProtocol: 'ws', sockPort: devServerPort },
-			exclude: [/node_module/, /outputCssLoader\.js/],
-		}),
+			new ReactRefreshWebpackPlugin({
+				overlay: {
+					sockHost: "127.0.0.1",
+					sockProtocol: "ws",
+					sockPort: devServerPort,
+				},
+				exclude: [/node_module/, /outputCssLoader\.js/],
+			}),
+		new SpriteLoaderPlugin(),
 	].filter(Boolean);
 };